restructuring to avoid run time loading complexities
